MINT-914 unblock ui in token form, decouple components, add placeholders for token (#454)
