+++ src/components/dialogs/tokens/TokenRowPlaceholder.scss
@@ -1,0 +1,29 @@
+@import "~styles/partials/variables";
+
+.root {
+  display: flex;
+  margin: 16px 8px;
+}
+
+.icon {
+  width: 64px;
+  height: 40px;
+  background-color: $color-grey400;
+}
+
+.info {
+  margin-left: 16px;
+}
+
+.name {
+  width: 100px;
+  height: 10px;
+  background-color: $color-grey300;
+  margin-bottom: 8px;
+}
+
+.balance {
+  width: 70px;
+  height: 10px;
+  background-color: $color-grey300;
+}
