+++ src/components/index.js
@@ -30,7 +30,7 @@ export { default as ModalDialog } from './dialogs/ModalDialog'
 export { default as CBEAddressDialog } from './dialogs/CBEAddressDialog'
 export { default as CBETokenDialog } from './dialogs/CBETokenDialog'
 export { default as OperationsSettingsDialog } from './dialogs/OperationsSettingsDialog'
-export { default as AddCurrencyDialog } from './dialogs/AddCurrencyDialog'
+export { default as AddCurrencyDialog } from './dialogs/tokens/AddCurrencyDialog'
 export { default as PollEditDialog } from './dialogs/poll/PollEditDialog/PollEditDialog'
 export { default as VoteDialog } from './dialogs/VoteDialog'
 export { default as PollDetailsDialog } from './dialogs/PollDetailsDialog'
